feat(vite): create a `virtual-mpa` plugin for dynamic pages discovery
